requirements.txt
Django~=4.2.5
uuid~=1.30
python-dotenv~=1.0.0
Pillow~=10.1.0